refactor: split main.js into feature modules
Extract four self-contained modules from main.js (~1238 → ~887 lines):
- app/features/windows-titlebar.js — custom in-page titlebar injection
- app/features/updater.js — auto-updater setup, events, manual check
- app/features/app-menu.js — application menu and About dialog
- app/features/staging-banner.js — staging build banner injection
